Move the test "show partitions from a datasource"
